@@ -22,14 +22,13 @@ module ActiveModelSerializers
           assert_equal serializable_resource.serializer_instance.attributes, {}
           assert_equal serializable_resource.serializer_instance.object, @resource
 
-          expected_errors_object =
-            { :errors =>
-              [
-                {
-                  source: { pointer: '/data/attributes/name' },
-                  detail: 'cannot be nil'
-                }
-              ]
+          expected_errors_object = {
+            :errors => [
+              {
+                source: { pointer: '/data/attributes/name' },
+                detail: 'cannot be nil'
+              }
+            ]
           }
           assert_equal serializable_resource.as_json, expected_errors_object
         end
@@ -48,13 +47,12 @@ module ActiveModelSerializers
           assert_equal serializable_resource.serializer_instance.attributes, {}
           assert_equal serializable_resource.serializer_instance.object, @resource
 
-          expected_errors_object =
-            { :errors =>
-              [
-                { :source => { :pointer => '/data/attributes/name' }, :detail => 'cannot be nil' },
-                { :source => { :pointer => '/data/attributes/name' }, :detail => 'must be longer' },
-                { :source => { :pointer => '/data/attributes/id' }, :detail => 'must be a uuid' }
-              ]
+          expected_errors_object = {
+            :errors => [
+              { :source => { :pointer => '/data/attributes/name' }, :detail => 'cannot be nil' },
+              { :source => { :pointer => '/data/attributes/name' }, :detail => 'must be longer' },
+              { :source => { :pointer => '/data/attributes/id' }, :detail => 'must be a uuid' }
+            ]
           }
           assert_equal serializable_resource.as_json, expected_errors_object
         end
