@@ -2,16 +2,15 @@ module ActiveModel
   class Serializer
     # This class hold all information about serializer's association.
     #
-    # @param [Symbol] name
-    # @param [ActiveModel::Serializer] serializer
-    # @param [Hash{Symbol => Object}] options
+    # @attr [Symbol] name
+    # @attr [ActiveModel::Serializer] serializer
+    # @attr [Hash{Symbol => Object}] options
     #
     # @example
     #  Association.new(:comments, CommentSummarySerializer)
     #
     Association = Struct.new(:name, :serializer, :options, :links, :meta) do
       # @return [Symbol]
-      #
       def key
         options.fetch(:key, name)
       end
