@@ -36,16 +36,10 @@ module ActiveModel::Serializer::Lint
     def test_read_attribute_for_serialization
       assert_respond_to resource, :read_attribute_for_serialization, 'The resource should respond to read_attribute_for_serialization'
       actual_arity = resource.method(:read_attribute_for_serialization).arity
-      if defined?(::Rubinius)
-        #  1 for def read_attribute_for_serialization(name); end
-        # -2 for alias :read_attribute_for_serialization :send for rbx because :shrug:
-        assert_includes [1, -2], actual_arity, "expected #{actual_arity.inspect} to be 1 or -2"
-      else
-        # using absolute value since arity is:
-        #  1 for def read_attribute_for_serialization(name); end
-        # -1 for alias :read_attribute_for_serialization :send
-        assert_includes [1, -1], actual_arity, "expected #{actual_arity.inspect} to be 1 or -1"
-      end
+      # using absolute value since arity is:
+      #  1 for def read_attribute_for_serialization(name); end
+      # -1 for alias :read_attribute_for_serialization :send
+      assert_equal 1, actual_arity.abs, "expected #{actual_arity.inspect}.abs to be 1 or -1"
     end
 
     # Passes if the object responds to <tt>as_json</tt> and if it takes
@@ -76,19 +70,15 @@ module ActiveModel::Serializer::Lint
       resource.to_json(nil)
     end
 
-    # Passes if the object responds to <tt>cache_key</tt> and if it takes no
-    # arguments (Rails 4.0) or a splat (Rails 4.1+).
+    # Passes if the object responds to <tt>cache_key</tt>
     # Fails otherwise.
     #
-    # <tt>cache_key</tt> returns a (self-expiring) unique key for the object, and
-    # is part of the (self-expiring) cache_key, which is used by the adapter.
-    # It is not required unless caching is enabled.
+    # <tt>cache_key</tt> returns a (self-expiring) unique key for the object,
+    # and is part of the (self-expiring) cache_key, which is used by the
+    # adapter. It is not required unless caching is enabled.
     def test_cache_key
       assert_respond_to resource, :cache_key
       actual_arity = resource.method(:cache_key).arity
-      # using absolute value since arity is:
-      #   0 for Rails 4.1+, *timestamp_names
-      #  -1 for Rails 4.0, no arguments
       assert_includes [-1, 0], actual_arity, "expected #{actual_arity.inspect} to be 0 or -1"
     end
 
