@@ -1,5 +1,5 @@
 module ActiveModel
   class Serializer
-    VERSION = '0.10.0.rc4'.freeze
+    VERSION = '0.10.1'.freeze
   end
 end
