@@ -8,10 +8,12 @@ module ActiveModelSerializers
         # Builds a JSON API Errors Object
         # {http://jsonapi.org/format/#errors JSON API Errors}
         #
-        # @param [ActiveModel::Serializer::ErrorSerializer]
-        # @return [Array<Symbol, Array<String>] i.e. attribute_name, [attribute_errors]
-        def self.resource_errors(error_serializer)
+        # @param [ActiveModel::Serializer::ErrorSerializer] error_serializer
+        # @return [Array<Symbol, Array<String>>] i.e. attribute_name, [attribute_errors]
+        def self.resource_errors(error_serializer, options)
           error_serializer.as_json.flat_map do |attribute_name, attribute_errors|
+            attribute_name = JsonApi.send(:transform_key_casing!, attribute_name,
+              options)
             attribute_error_objects(attribute_name, attribute_errors)
           end
         end
