@@ -1,8 +1,7 @@
 module ActiveModelSerializers
   module Adapter
     class Null < Base
-      # Since options param is not being used, underscored naming of the param
-      def serializable_hash(_options = nil)
+      def serializable_hash(*)
         {}
       end
     end
