@@ -1,3 +1,4 @@
+require 'active_support/core_ext/array/extract_options'
 module ActiveModelSerializers
   class SerializationContext
     class << self
@@ -22,12 +23,17 @@ module ActiveModelSerializers
 
     attr_reader :request_url, :query_parameters, :key_transform
 
-    def initialize(request, options = {})
-      @request_url = request.original_url[/\A[^?]+/]
-      @query_parameters = request.query_parameters
+    def initialize(*args)
+      options = args.extract_options!
+      if args.size == 1
+        request = args.pop
+        options[:request_url] = request.original_url[/\A[^?]+/]
+        options[:query_parameters] = request.query_parameters
+      end
+      @request_url = options.delete(:request_url)
+      @query_parameters = options.delete(:query_parameters)
       @url_helpers = options.delete(:url_helpers) || self.class.url_helpers
       @default_url_options = options.delete(:default_url_options) || self.class.default_url_options
-      @key_transform = options.delete(:key_transform)
     end
   end
 end
