@@ -2,27 +2,46 @@ module ActiveModelSerializers
   module Test
     module Schema
       # A Minitest Assertion that test the response is valid against a schema.
-      # @params schema_path [String] a custom schema path
-      # @params message [String] a custom error message
+      # @param schema_path [String] a custom schema path
+      # @param message [String] a custom error message
       # @return [Boolean] true when the response is valid
       # @return [Minitest::Assertion] when the response is invalid
       # @example
       #   get :index
       #   assert_response_schema
       def assert_response_schema(schema_path = nil, message = nil)
-        matcher = AssertResponseSchema.new(schema_path, response, message)
+        matcher = AssertResponseSchema.new(schema_path, request, response, message)
         assert(matcher.call, matcher.message)
       end
 
-      MissingSchema = Class.new(Errno::ENOENT)
-      InvalidSchemaError = Class.new(StandardError)
+      def assert_request_schema(schema_path = nil, message = nil)
+        matcher = AssertRequestSchema.new(schema_path, request, response, message)
+        assert(matcher.call, matcher.message)
+      end
 
-      class AssertResponseSchema
-        attr_reader :schema_path, :response, :message
+      # May be renamed
+      def assert_request_response_schema(schema_path = nil, message = nil)
+        assert_request_schema(schema_path, message)
+        assert_response_schema(schema_path, message)
+      end
 
-        def initialize(schema_path, response, message)
+      def assert_schema(payload, schema_path = nil, message = nil)
+        matcher = AssertSchema.new(schema_path, request, response, message, payload)
+        assert(matcher.call, matcher.message)
+      end
+
+      MissingSchema = Class.new(Minitest::Assertion)
+      InvalidSchemaError = Class.new(Minitest::Assertion)
+
+      class AssertSchema
+        attr_reader :schema_path, :request, :response, :message, :payload
+
+        # Interface may change.
+        def initialize(schema_path, request, response, message, payload = nil)
           require_json_schema!
+          @request = request
           @response = response
+          @payload = payload
           @schema_path = schema_path || schema_path_default
           @message = message
           @document_store = JsonSchema::DocumentStore.new
@@ -32,7 +51,7 @@ module ActiveModelSerializers
         def call
           json_schema.expand_references!(store: document_store)
           status, errors = json_schema.validate(response_body)
-          @message ||= errors.map(&:to_s).to_sentence
+          @message = [message, errors.map(&:to_s).to_sentence].compact.join(': ')
           status
         end
 
@@ -41,11 +60,11 @@ module ActiveModelSerializers
         attr_reader :document_store
 
         def controller_path
-          response.request.filtered_parameters[:controller]
+          request.filtered_parameters[:controller]
         end
 
         def action
-          response.request.filtered_parameters[:action]
+          request.filtered_parameters[:action]
         end
 
         def schema_directory
@@ -66,6 +85,10 @@ module ActiveModelSerializers
 
         def response_body
           load_json(response.body)
+        end
+
+        def request_params
+          request.env['action_dispatch.request.request_parameters']
         end
 
         def json_schema
@@ -96,6 +119,18 @@ module ActiveModelSerializers
           require 'json_schema'
         rescue LoadError
           raise LoadError, "You don't have json_schema installed in your application. Please add it to your Gemfile and run bundle install"
+        end
+      end
+      class AssertResponseSchema < AssertSchema
+        def initialize(*)
+          super
+          @payload = response_body
+        end
+      end
+      class AssertRequestSchema < AssertSchema
+        def initialize(*)
+          super
+          @payload = request_params
         end
       end
     end
