@@ -8,9 +8,11 @@ module ActiveModelSerializers
           context = Minitest::Mock.new
           context.expect(:request_url, URI)
           context.expect(:query_parameters, {})
-          context.expect(:key_transform, key_transform)
-          @options = {}
-          @options[:serialization_context] = context
+          options = {}
+          options[:key_transform] = key_transform if key_transform
+          options[:serialization_context] = context
+          serializer = CustomBlogSerializer.new(@blog)
+          @adapter = ActiveModelSerializers::Adapter::Json.new(serializer, options)
         end
 
         Post = Class.new(::Model)
@@ -18,74 +20,72 @@ module ActiveModelSerializers
           attributes :id, :title, :body, :publish_at
         end
 
-        def setup
+        setup do
           ActionController::Base.cache_store.clear
           @blog = Blog.new(id: 1, name: 'My Blog!!', special_attribute: 'neat')
-          serializer = CustomBlogSerializer.new(@blog)
-          @adapter = ActiveModelSerializers::Adapter::Json.new(serializer)
         end
 
-        def test_key_transform_default
+        def test_transform_default
           mock_request
           assert_equal({
             blog: { id: 1, special_attribute: 'neat', articles: nil }
-          }, @adapter.serializable_hash(@options))
+          }, @adapter.serializable_hash)
         end
 
-        def test_key_transform_global_config
+        def test_transform_global_config
           mock_request
           result = with_config(key_transform: :camel_lower) do
-            @adapter.serializable_hash(@options)
+            @adapter.serializable_hash
           end
           assert_equal({
             blog: { id: 1, specialAttribute: 'neat', articles: nil }
           }, result)
         end
 
-        def test_key_transform_serialization_ctx_overrides_global_config
+        def test_transform_serialization_ctx_overrides_global_config
           mock_request(:camel)
           result = with_config(key_transform: :camel_lower) do
-            @adapter.serializable_hash(@options)
+            @adapter.serializable_hash
           end
           assert_equal({
             Blog: { Id: 1, SpecialAttribute: 'neat', Articles: nil }
           }, result)
         end
 
-        def test_key_transform_undefined
+        def test_transform_undefined
           mock_request(:blam)
           result = nil
           assert_raises NoMethodError do
-            result = @adapter.serializable_hash(@options)
+            result = @adapter.serializable_hash
           end
         end
 
-        def test_key_transform_dashed
-          mock_request(:dashed)
+        def test_transform_dash
+          mock_request(:dash)
           assert_equal({
             blog: { id: 1, :"special-attribute" => 'neat', articles: nil }
-          }, @adapter.serializable_hash(@options))
+          }, @adapter.serializable_hash)
         end
 
-        def test_key_transform_unaltered
+        def test_transform_unaltered
           mock_request(:unaltered)
           assert_equal({
             blog: { id: 1, special_attribute: 'neat', articles: nil }
-          }, @adapter.serializable_hash(@options))
+          }, @adapter.serializable_hash)
         end
 
-        def test_key_transform_camel
+        def test_transform_camel
           mock_request(:camel)
           assert_equal({
             Blog: { Id: 1, SpecialAttribute: 'neat', Articles: nil }
-          }, @adapter.serializable_hash(@options))
+          }, @adapter.serializable_hash)
         end
 
-        def test_key_transform_camel_lower
+        def test_transform_camel_lower
           mock_request(:camel_lower)
           assert_equal({
             blog: { id: 1, specialAttribute: 'neat', articles: nil }
-          }, @adapter.serializable_hash(@options))
+          }, @adapter.serializable_hash)
         end
       end
     end
