@@ -58,9 +58,10 @@ module ActiveModelSerializers
 
         def test_limiting_fields
           actual = ActiveModelSerializers::SerializableResource.new(
-            [@first_post, @second_post], adapter: :json_api,
-                                         fields: { posts: %w(title comments blog author) })
-                                                               .serializable_hash
+            [@first_post, @second_post],
+            adapter: :json_api,
+            fields: { posts: %w(title comments blog author) }
+          ).serializable_hash
           expected = [
             {
               id: '1',
