@@ -31,8 +31,20 @@ class ApiAssertion
     get("/caching/#{on_off}")
   end
 
+  def get_fragment_caching(on_off = 'on'.freeze)
+    get("/fragment_caching/#{on_off}")
+  end
+
   def get_non_caching(on_off = 'on'.freeze)
     get("/non_caching/#{on_off}")
+  end
+
+  def debug(msg = '')
+    if block_given? && ENV['DEBUG'] =~ /\Atrue|on|0\z/i
+      STDERR.puts yield
+    else
+      STDERR.puts msg
+    end
   end
 
   private
@@ -54,24 +66,25 @@ class ApiAssertion
   def expected
     @expected ||=
       {
-        'post' =>  {
-          'id' =>  1337,
-          'title' => 'New Post',
+        'primary_resource' => {
+          'id' => 1337,
+          'title' => 'New PrimaryResource',
           'body' =>  'Body',
-          'comments' => [
+          'virtual_attribute' => {
+            'id' => 999,
+            'name' => 'Free-Range Virtual Attribute'
+          },
+          'has_one_relationship' => {
+            'id' => 42,
+            'first_name' => 'Joao',
+            'last_name' => 'Moura'
+          },
+          'has_many_relationships' => [
             {
               'id' => 1,
-              'body' => 'ZOMG A COMMENT'
+              'body' => 'ZOMG A HAS MANY RELATIONSHIP'
             }
-          ],
-          'blog' =>  {
-            'id' =>  999,
-            'name' => 'Custom blog'
-          },
-          'author' => {
-            'id' => 42,
-            'name' => 'Joao Moura.'
-          }
+          ]
         }
     }
   end
@@ -84,33 +97,23 @@ class ApiAssertion
       STDERR.puts message unless ENV['SUMMARIZE']
     end
   end
-
-  def debug(msg = '')
-    if block_given? && ENV['DEBUG'] =~ /\Atrue|on|0\z/i
-      STDERR.puts yield
-    else
-      STDERR.puts msg
-    end
-  end
 end
 assertion = ApiAssertion.new
 assertion.valid?
-# STDERR.puts assertion.get_status
+assertion.debug { assertion.get_status }
 
 time = 10
 {
   'caching on: caching serializers: gc off' => { disable_gc: true, send: [:get_caching, 'on'] },
-  # 'caching on: caching serializers: gc on' => { disable_gc: false, send: [:get_caching, 'on'] },
-  'caching off: caching serializers: gc off' => { disable_gc: true, send: [:get_caching, 'off'] },
-  # 'caching off: caching serializers: gc on' => { disable_gc: false, send: [:get_caching, 'off'] },
+  'caching on: fragment caching serializers: gc off' => { disable_gc: true, send: [:get_fragment_caching, 'on'] },
   'caching on: non-caching serializers: gc off' => { disable_gc: true, send: [:get_non_caching, 'on'] },
-  # 'caching on: non-caching serializers: gc on' => { disable_gc: false, send: [:get_non_caching, 'on'] },
+  'caching off: caching serializers: gc off' => { disable_gc: true, send: [:get_caching, 'off'] },
+  'caching off: fragment caching serializers: gc off' => { disable_gc: true, send: [:get_fragment_caching, 'off'] },
   'caching off: non-caching serializers: gc off' => { disable_gc: true, send: [:get_non_caching, 'off'] }
-  # 'caching off: non-caching serializers: gc on' => { disable_gc: false, send: [:get_non_caching, 'off'] }
 }.each do |label, options|
   assertion.clear
   Benchmark.ams(label, time: time, disable_gc: options[:disable_gc]) do
     assertion.send(*options[:send])
   end
-  # STDERR.puts assertion.get_status(options[:send][-1])
+  assertion.debug { assertion.get_status(options[:send][-1]) }
 end
